fix open in etherscan link
