+++ apps/remix-ide/src/app/udapp/make-udapp.js
@@ -1,31 +1,9 @@
 import Registry from '../state/registry'
 
 var remixLib = require('@remix-project/remix-lib')
-var yo = require('yo-yo')
 var EventsDecoder = remixLib.execution.EventsDecoder
 
-const transactionDetailsLinks = {
-  Main: 'https://www.etherscan.io/tx/',
-  Rinkeby: 'https://rinkeby.etherscan.io/tx/',
-  Ropsten: 'https://ropsten.etherscan.io/tx/',
-  Kovan: 'https://kovan.etherscan.io/tx/',
-  Goerli: 'https://goerli.etherscan.io/tx/'
-}
-
-function txDetailsLink (network, hash) {
-  if (transactionDetailsLinks[network]) {
-    return transactionDetailsLinks[network] + hash
-  }
-}
-
 export function makeUdapp (blockchain, compilersArtefacts, logHtmlCallback) {
-  // ----------------- UniversalDApp -----------------
-  // TODO: to remove when possible
-  blockchain.event.register('transactionBroadcasted', (txhash, networkName) => {
-    var txLink = txDetailsLink(networkName, txhash)
-    if (txLink && logHtmlCallback) logHtmlCallback(yo`<a href="${txLink}" target="_blank">${txLink}</a>`)
-  })
-
   // ----------------- Tx listener -----------------
   const _transactionReceipts = {}
   const transactionReceiptResolver = (tx, cb) => {
